4.16.0.74-old1: experimental release by Hugh on 2015-10-15 14:30 (email Oct 15, 2015 at 7:19 PM) ftp.cs.toronto.edu:/pub/hugh/jove-dev/experimental/jove4.16.0.74-old1.tgz tarball sha256sum ca679b63c38b424eed3e16d3a2ab40e4463c70bbc16e83164ae4ad692de26da3
Changes:
- minor portability improvements
- new jove.spec for Fedora and EPEL
- fill-comment now handles paragraphs
- fill-region can be undone via yank-pop
- support for xterm-256color (same as xterm

+++ abbrev.c
@@ -138,7 +138,7 @@ AbbrevExpand()
 	if ((ap = lookup_abbrev(A_tables[curbuf->b_major], wordbuf)) != NULL
 	|| (ap = lookup_abbrev(A_tables[GLOBAL], wordbuf)) != NULL)
 	{
-		del_char(BACKWARD, (wp - wordbuf), NO);
+		del_char(BACKWARD, (int)(wp - wordbuf), NO);
 
 		for (cp = ap->a_phrase; (c = *cp) != '\0'; ) {
 			if (UC_count > 0 && jislower(c)
